Worked with conflicts and added Future Builder
